facets repository reactive refactoring
