+++ odd-platform-api/src/main/java/org/opendatadiscovery/oddplatform/service/SearchServiceImpl.java
@@ -27,7 +27,8 @@ import org.opendatadiscovery.oddplatform.mapper.FacetStateMapper;
 import org.opendatadiscovery.oddplatform.mapper.SearchMapper;
 import org.opendatadiscovery.oddplatform.model.tables.pojos.SearchFacetsPojo;
 import org.opendatadiscovery.oddplatform.repository.DataEntityRepository;
-import org.opendatadiscovery.oddplatform.repository.SearchFacetRepository;
+import org.opendatadiscovery.oddplatform.repository.reactive.ReactiveDataEntityRepository;
+import org.opendatadiscovery.oddplatform.repository.reactive.ReactiveSearchFacetRepository;
 import org.opendatadiscovery.oddplatform.utils.Page;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Flux;
@@ -45,9 +46,9 @@ public class SearchServiceImpl implements SearchService {
     private final SearchMapper searchMapper;
     private final FacetStateMapper facetStateMapper;
     private final DataEntityMapper dataEntityMapper;
-    private final SearchFacetRepository searchFacetRepository;
+    private final ReactiveSearchFacetRepository searchFacetRepository;
     private final DataEntityRepository dataEntityRepository;
-
+    private final ReactiveDataEntityRepository reactiveDataEntityRepository;
     private final AuthIdentityProvider authIdentityProvider;
 
     @Override
@@ -58,26 +59,20 @@ public class SearchServiceImpl implements SearchService {
                                                         final String query) {
         return fetchFacetState(searchId)
             .map(facetStateMapper::pojoToState)
-            .flatMapIterable(state -> {
-                final Set<Long> selectedFilters = state.getFacetEntities(FacetType.lookup(facetType.getValue()))
-                    .stream()
-                    .filter(SearchFilterDto::isSelected)
-                    .map(SearchFilterDto::getEntityId)
-                    .collect(Collectors.toSet());
-
-                return createFacetFetchOperation(facetType, query, page, size)
-                    .apply(state)
-                    .entrySet().stream()
-                    .filter(e -> !selectedFilters.contains(e.getKey().getEntityId()))
-                    .map(e -> searchMapper.mapCountableSearchFilter(e.getKey(), e.getValue()))
-                    .sorted(Comparator.comparing(CountableSearchFilter::getCount).reversed())
-                    .collect(Collectors.toList());
-            });
+            .flatMapMany(state -> getFacetFetchOperation(facetType, query, page, size)
+                .apply(state)
+                .map(facetMap -> removeSelectedFiltersAndMap(facetMap, state, facetType))
+                .flatMapIterable(Function.identity())
+            );
     }
 
     @Override
     public Mono<SearchFacetsData> getFacets(final UUID searchId) {
-        return fetchFacetState(searchId).flatMap(p -> getFacetsData(p.getId(), facetStateMapper.pojoToState(p)));
+        return fetchFacetState(searchId)
+            .flatMap(pojo -> {
+                final FacetStateDto state = facetStateMapper.pojoToState(pojo);
+                return getFacetsData(pojo.getId(), state);
+            });
     }
 
     @Override
@@ -86,7 +81,7 @@ public class SearchServiceImpl implements SearchService {
 
         return Mono
             .just(facetStateMapper.mapStateToPojo(state))
-            .map(searchFacetRepository::persistFacetState)
+            .flatMap(searchFacetRepository::create)
             .flatMap(p -> getFacetsData(p.getId(), state));
     }
 
@@ -95,10 +90,12 @@ public class SearchServiceImpl implements SearchService {
         final FacetStateDto delta = facetStateMapper.mapForm(formData);
 
         return fetchFacetState(searchId)
-            .map(facetStateMapper::pojoToState)
-            .map(currentState -> FacetStateDto.merge(currentState, delta))
-            .map(mergedState -> facetStateMapper.mapStateToPojo(searchId, mergedState))
-            .map(searchFacetRepository::updateFacetState)
+            .map(facetsRecord -> {
+                final FacetStateDto currentState = facetStateMapper.pojoToState(facetsRecord);
+                final FacetStateDto mergedState = FacetStateDto.merge(currentState, delta);
+                return facetStateMapper.mapStateToPojo(searchId, mergedState);
+            })
+            .flatMap(searchFacetRepository::update)
             .flatMap(p -> getFacetsData(p.getId(), facetStateMapper.pojoToState(p)));
     }
 
@@ -121,20 +118,21 @@ public class SearchServiceImpl implements SearchService {
     }
 
     @Override
-    public Flux<DataEntityRef> getQuerySuggestions(final String query, final Integer entityClassId,
+    public Flux<DataEntityRef> getQuerySuggestions(final String query,
+                                                   final Integer entityClassId,
                                                    final Boolean manuallyCreated) {
-        return Flux
-            .fromIterable(dataEntityRepository.getQuerySuggestions(query, entityClassId, manuallyCreated))
+        return reactiveDataEntityRepository.getQuerySuggestions(query, entityClassId, manuallyCreated)
             .map(dataEntityMapper::mapRef);
     }
 
     private Mono<SearchFacetsData> getFacetsData(final UUID searchId, final FacetStateDto state) {
-        final Mono<Map<SearchFilterId, Long>> entityClassFacet = fetchEntityClassFacet(state);
+        final Mono<Map<SearchFilterId, Long>> entityClassFacet = searchFacetRepository
+            .getEntityClassFacetForDataEntity(state);
 
-        final Mono<Long> allCount = Mono.fromCallable(() -> dataEntityRepository.countByState(state));
+        final Mono<Long> allCount = reactiveDataEntityRepository.countByState(state);
 
         final Mono<Long> myObjectsCount = authIdentityProvider.fetchAssociatedOwner()
-            .map(owner -> dataEntityRepository.countByState(state, owner))
+            .flatMap(owner -> reactiveDataEntityRepository.countByState(state, owner))
             .switchIfEmpty(Mono.just(0L));
 
         return Mono.zip(entityClassFacet, allCount, myObjectsCount).map(
@@ -162,29 +160,42 @@ public class SearchServiceImpl implements SearchService {
             }));
     }
 
-    private Mono<Map<SearchFilterId, Long>> fetchEntityClassFacet(final FacetStateDto state) {
-        return Mono.fromCallable(() -> searchFacetRepository.getEntityClassFacet(state));
-    }
-
     private Mono<SearchFacetsPojo> fetchFacetState(final UUID searchId) {
-        return Mono
-            .fromCallable(() -> searchFacetRepository.getFacetState(searchId))
-            .flatMap(optional -> optional.isEmpty()
-                ? Mono.error(new NotFoundException("Search not found"))
-                : Mono.just(optional.get()));
+        return searchFacetRepository.get(searchId)
+            .switchIfEmpty(Mono.error(() -> new NotFoundException("Search not found")));
     }
 
-    private Function<FacetStateDto, Map<SearchFilterId, Long>> createFacetFetchOperation(
+    private Function<FacetStateDto, Mono<Map<SearchFilterId, Long>>> getFacetFetchOperation(
         final MultipleFacetType facetType,
         final String query,
         final Integer page,
         final Integer size
     ) {
         return switch (facetType) {
-            case TAGS -> s -> searchFacetRepository.getTagFacet(query, page, size, s);
-            case OWNERS -> s -> searchFacetRepository.getOwnerFacet(query, page, size, s);
-            case TYPES -> s -> searchFacetRepository.getTypeFacet(query, page, size, s);
-            case GROUPS -> s -> searchFacetRepository.getGroupFacet(query, page, size, s);
+            case TAGS -> s -> searchFacetRepository.getTagFacetForDataEntity(query, page, size, s);
+            case OWNERS -> s -> searchFacetRepository.getOwnerFacetForDataEntity(query, page, size, s);
+            case TYPES -> s -> searchFacetRepository.getTypeFacetForDataEntity(query, page, size, s);
+            case GROUPS -> s -> searchFacetRepository.getGroupFacetForDataEntity(query, page, size, s);
         };
+    }
+
+    private List<CountableSearchFilter> removeSelectedFiltersAndMap(final Map<SearchFilterId, Long> facetFiltersMap,
+                                                                    final FacetStateDto state,
+                                                                    final MultipleFacetType facetType) {
+        final Set<Long> selectedFilters = getSelectedFilters(state, facetType);
+        return facetFiltersMap.entrySet().stream()
+            .filter(e -> !selectedFilters.contains(e.getKey().getEntityId()))
+            .map(e -> searchMapper.mapCountableSearchFilter(e.getKey(), e.getValue()))
+            .sorted(Comparator.comparing(CountableSearchFilter::getCount).reversed())
+            .collect(Collectors.toList());
+    }
+
+    private Set<Long> getSelectedFilters(final FacetStateDto state,
+                                         final MultipleFacetType facetType) {
+        return state.getFacetEntities(FacetType.lookup(facetType.getValue()))
+            .stream()
+            .filter(SearchFilterDto::isSelected)
+            .map(SearchFilterDto::getEntityId)
+            .collect(Collectors.toSet());
     }
 }
